Out with the old, in with the new
This commit is long overdue. The main change is to upgrade shoulda's
hard dependency on shoulda-matchers from 2.x to 3.x. Since this by
itself is a pretty consequential change, though, I also took the
opportunity to make some more at the same time that I think are
reasonable:

* Remove support for Rails 3, upgrading dependency to 4, modifying
  Appraisals file and fixing Cucumber specs to match.
* Remove support for RSpec, along with related Cucumber specs. People
  who use RSpec won't be using the `shoulda` gem -- they'll be using
  `shoulda-matchers` instead. There's already a good collection of
  integration specs there, so we don't need the tests here.
* Remove support for Ruby 1.9, along with the entry in .travis.yml.

Due to the backward-incompatible nature of these changes, my thought is
to release a version 4 of shoulda that includes them.